requirements.txt
Flask==3.1.3
pandas==2.3.3
requests==2.32.5
python-dotenv==1.2.2
anthropic==0.84.0
spotipy==2.26.0
numpy==2.4.3